ai_agent_book/topic_1.4/requirements.txt
sentence-transformers
torch
openai
python-dotenv
requests
numpy
scikit-learn
faiss-cpu
ipykernel
rich
markdown